bugfix: UNINSTALL_KEEP_DATA setting was not recognized in wordpress
git-svn-id: https://plugins.svn.wordpress.org/extrawatch-pro/trunk@733275 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ components/com_extrawatch/src/env/wordpress/class.extrawatch.setup.wordpress.php
@@ -5,7 +5,7 @@
  * ExtraWatch - A real-time ajax monitor and live stats
  * @package ExtraWatch
  * @version 2.1
- * @revision 793
+ * @revision 794
  * @license http://www.gnu.org/licenses/gpl-3.0.txt     GNU General Public License v3
  * @copyright (C) 2013 by CodeGravity.com - All rights reserved!
  * @website http://www.extrawatch.com
@@ -88,6 +88,18 @@ class ExtraWatchSetupWordpress implements ExtraWatchSetup
 
   static function drop_tables($database)
   {
+
+      try {
+          $database->setQuery("SELECT `value` as keepData from #__extrawatch_config where `name` = 'EXTRAWATCH_UNINSTALL_KEEP_DATA' limit 1 ");
+          $keepData = @$database->loadResult();
+      } catch (Exception $e) {
+          // suppress
+      }
+
+    if (@$keepData) {
+        echo("<b>Not deleting ExtraWatch database tables, because you set KEEP_DATA in Settings to true. <br/>Do not forget to delete these tables later, or install new version of ExtraWatch</b> <br/><br/>");
+        return;
+    }
 
     $query = "DROP TABLE #__extrawatch";
     $database->setQuery(trim($database->replaceDbPrefix($query)));
